Add Item Shop (with Sell Function).
Add buying and selling prices to item ScriptableObjects.
Add TMP Sprite Asset for gold, to allow gold sprites in TMP_Text objects.
Refactor AmtConfirm Window to be general class for any actions that require the player to enter and confirm an amount.

Player is now able to sell items at the shop.